remove integrations using SASL and fix keystore test
