--- integrationtests/security-it/src/test/resources/jgroups-tcp-sasl-prop-handler-node0.xml
@@ -1,62 +1,0 @@
-<config xmlns="urn:org:jgroups"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="urn:org:jgroups http://www.jgroups.org/schema/jgroups-5.2.xsd">
-
-   <TCP  bind_addr="${jgroups.bind.address,jgroups.tcp.address:127.0.0.1}"
-         bind_port="7800" port_range="10"
-         recv_buf_size="20000000"
-         send_buf_size="640000"
-         bundler_type="transfer-queue"
-         diag.enabled="true"
-         thread_naming_pattern="cl"
-
-         thread_pool.min_threads="0"
-         thread_pool.max_threads="8"
-         thread_pool.keep_alive_time="5000"
-
-         thread_pool.thread_dumps_threshold="${jgroups.thread_dumps_threshold:10000}"
-   />
-   <RED/>
-
-   <MPING mcast_addr="${jgroups.mping.mcast_addr:228.2.4.6}"
-          mcast_port="${jgroups.mping.mcast_port:43366}"
-          break_on_coord_rsp="true"
-          num_discovery_runs="3"
-          ip_ttl="${jgroups.udp.ip_ttl:2}"/>
-
-   <MERGE3/>
-
-   <FD_SOCK/>
-   <!-- Suspect node `timeout` to `timeout + timeout_check_interval` millis after the last heartbeat -->
-   <FD_ALL3 timeout="10000"
-           interval="2000"
-   />
-   <VERIFY_SUSPECT timeout="1000"/>
-
-   <pbcast.NAKACK2
-                    xmit_interval="200"
-                    xmit_table_num_rows="50"
-                    xmit_table_msgs_per_row="1024"
-                    xmit_table_max_compaction_time="30000"/>
-   <UNICAST3
-              conn_close_timeout="5000"
-              xmit_interval="200"
-              xmit_table_num_rows="50"
-              xmit_table_msgs_per_row="1024"
-              xmit_table_max_compaction_time="30000"
-   />
-   <SASL mech="DIGEST-MD5"
-         server_callback_handler_class="org.jgroups.auth.sasl.SimpleAuthorizingCallbackHandler"
-         client_callback_handler_class="org.jgroups.auth.sasl.SimpleAuthorizingCallbackHandler"
-         sasl_props="com.sun.security.sasl.digest.realm=test_realm" />
-
-   <pbcast.STABLE desired_avg_gossip="5000"
-                  max_bytes="1M"
-   />
-   <pbcast.GMS print_local_addr="false"
-               join_timeout="${jgroups.join_timeout:2000}"/>
-
-   <UFC max_credits="4m" min_threshold="0.40"/>
-   <MFC max_credits="4m" min_threshold="0.40"/>
-   <FRAG3/>
-</config>
